requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.2
google-generativeai==0.7.2
Pillow==10.3.0
lxml==5.2.2
tqdm==4.66.4
